requirements.txt
gdown>=5.2.0
gradio>=5.33.0
ipykernel
jupyter
numpy>=2.2.6
opencv-python-headless>=4.11.0.86
pillow>=11.2.1
pyprojroot>=0.3.0
python-box==6.0.2
PyYAML>=6.0.2
torch>=2.7.1
torchvision>=0.22.1

-e .